+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Chg #207: Replace two `RouteCollectorInterface` methods `addRoute()` and `addGroup()` to single `addRoute()` (@vjik)
 - Chg #222: Make `Route`, `Group` and `MatchingResult` dispatcher-independent (@rustamwin, @vjik)
 - Enh #229: Add URL arguments' psalm type in `UrlGeneratorInterface` (@vjik)
+- Chg #256: Bump minimum PHP version to 8.1 (@vjik)
+- Enh #256: Mark readonly properties (@vjik)
 
 ## 3.1.0 February 20, 2024